bot_tg/requirements.txt
bs4==0.0.1
pyTelegramBotAPI==4.13.0
python-decouple==3.8
requests==2.31.0
